+++ CHANGES.rst
@@ -5,6 +5,7 @@ Changelog
 ----------------
 
 - Prevent builtins shadowing in DTML Var.
+  (`Zope#1285 <https://github.com/zopefoundation/Zope/issues/1285>`_)
 
 
 5.1 (2026-01-19)